Check for if muted role is none
The muted role could return none in this context if the role was deleted and the bot gets run again. The role ID would no longer point to a role therefore will not work for the rest of the function. This check ensures that the role will always exist before we try to use it.