Added a dynamic class loading system for commands
Added a dynamic class loader system for commands. This allows the administrator of the bot to add/remove commands without needing to restart the bot. It's also good to show my python skills ;)
Added a reload command to reload all other commands (And it reloads itself! reload-ception!)
Updated all the commands to use the new command registry system and the dynamic class loading!