Fixed a bug when checking cmd and event handlers
Fixed a bug when checking the subclass of an event or command handler.
This was because the event handler or command base were being deleted
from the system modules when loading stuff which caused it to not have
a valid reference.

Whats even sadder is that I wrapped them with `type()` initially which
doesn't work for this comparison and meant this code had a bug causing
a crash when you had a class that wasn't properly setup to handle a
command or event.